+++ plugins/websocket_server/WebsocketServer.hh
@@ -37,7 +37,45 @@ namespace ignition
     ///
     /// # Websocket Server Interface
     ///
-    /// \todo: Describe the websocket API.
+    /// The websocket server listens for incoming requests and sends
+    /// messages based on the requests.
+    ///
+    /// All messages on the websocket, incoming and outgoing, are structured
+    /// in a frame that consists of four comma separated components:
+    ///     1. `operation`: string,
+    ///     2. `topic_name`: string,
+    ///     3. `message_type`: string, and
+    ///     4. `payload`: serialized data.
+    ///
+    /// The `operation` component is mandatory and must be one of:
+    ///     1. "sub": Subscribe to the topic in the `topic_name` component,
+    ///     2. "pub": Publish a message from the Ingition Transport topic in
+    ///               the `topic_name` component,
+    ///     3. "topics": Get the list of available topics, and
+    ///     4. "protos": Get a string containing all the protobuf
+    ///                  definitions.
+    ///
+    /// The `topic_name` component is mandatory for the "sub" and "pub"
+    /// operations. If present, it must be the name of an Ignition Transport
+    /// topic.
+    ///
+    /// The `message_type` component is mandatory for the "pub" operation. If
+    /// present it names the Ignition Message type, such as
+    /// "ignition.msgs.Clock".
+    ///
+    /// The `payload` component is mandatory for the "pub" operation. If
+    /// present, it contains a serialized string of an Igntion Message.
+    ///
+    /// ## Example frames
+    ///
+    /// 1. Get the list of topics: `topics,,,`
+    ///
+    /// 2. Get the protobuf definitions: `protos,,,`
+    ///
+    /// 3. Subscribe to the "/clock" topic: `sub,/clock,,`
+    ///
+    /// 4. Websocker server publishing data on the "/clock" topic:
+    ///    `pub,/clock,ignition.msgs.Clock,<serialized_data>`
     ///
     /// # Example usage
     ///
@@ -80,10 +118,10 @@ namespace ignition
 
       public: void OnConnect(int _socketId);
       public: void OnDisconnect(int _socketId);
+
       public: void OnMessage(int _socketId, const std::string &_msg);
 
       public: void OnRequestMessage(int _socketId, const std::string &_msg);
-
 
       private: ignition::transport::Node node;
 
@@ -104,7 +142,17 @@ namespace ignition
                    const char *_data, const size_t _size);
 
       public: std::mutex mutex;
+
+      /// \brief A mutex used in the OnWebsocketSubscribedMessage
+      /// function.
+      public: std::mutex subscriptionMutex;
+
+      /// \brief All of the websocket connections.
       public: std::map<int, std::unique_ptr<Connection>> connections;
+
+      /// \brief All of the subscribed Ignition topics.
+      /// The key is the topic name, and the value is the set of websocket
+      /// connections that have subscribed to the topic.
       public: std::map<std::string, std::set<int>> topicConnections;
 
       /// \brief Time of last publication for each subscribed topic. The key
@@ -113,6 +161,33 @@ namespace ignition
       private: std::map<std::string,
                std::chrono::time_point<std::chrono::steady_clock>>
                  topicTimestamps;
+
+      /// \brief The set of valid operations. This enum must align with the
+      /// `operations` member variable.
+      private: enum Operation
+               {
+                 /// \brief Subscribe to a topic.
+                 SUBSCRIBE = 0,
+
+                 /// \brief Publish a message from a topic.
+                 PUBLISH = 1,
+
+                 /// \brief Get the list of topics.
+                 TOPICS = 2,
+
+                 /// \brief Get the protobuf definitions.
+                 PROTOS = 3,
+               };
+
+      /// \brief The set of valid operations, in string  form. These values
+      /// can be sent in websocket message frames.
+      /// These valus must align with the `Operation` enum.
+      private: std::vector<std::string> operations{
+                 "sub", "pub", "topics", "protos"};
+
+      /// \brief Store publish headers for topics. This is here to improve
+      /// performance. Keys are topic names and values are frame headers.
+      private: std::map<std::string, std::string> publishHeaders;
 
       /// \brief Period at which messages will be published on the websocket
       /// for each subscribed topic.