+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ### Ignition Launch 1.X.X
 
+1. Changed the websocket messages to be more efficient when transmitting
+   protobuf messages.
+    * [Pull Request 22](https://github.com/ignitionrobotics/ign-launch/pull/22)
+
 1. Added ability to get protobuf definitions from the websocket server
     * [Pull Request 21](https://github.com/ignitionrobotics/ign-launch/pull/21)